More refactoring in preparation for GESIS.
